shipinit bonk damage
